+++ dkg-gadget/src/storage/mod.rs
@@ -1,2 +1,4 @@
+pub mod clear;
 pub mod misbehaviour_reports;
+pub mod proposals;
 pub mod public_keys;
